Unify JFR ID recomputations for packages, modules and classloaders
